FInal Submission
I have completed the requirements as well as the extra credit.

Followed and MVC standard pattern with the root controller being a UIViewSlitController managing both the ContactList and CreateContact screens, which each have their own controllers and subviews. I think this would be a very extendable and scalable solution after a bit of cleaning up.